Make grabix compile again
As proposed in https://github.com/arq5x/grabix/issues/38.

'\0' is still the character terminating strings in C, but line->s is a pointer to a string, not the string itself. I added the cast to make the semantics clearer. On a sidenote, the kstring gets dynamically allocated but never freed. This is small memory leak. More importantly, I am a bit uncertain about who is freeing line->s . So, should there ever be an issue with memory, this may be worth a second look.